Moving the .Net Usage type to the specific test that uses it
